Refactored code + restructured files
