--- .vscode/settings.json
@@ -1,8 +1,0 @@
-{
-  "files.exclude": {
-    "**/.classpath": true,
-    "**/.project": true,
-    "**/.settings": true,
-    "**/.factorypath": true
-  }
-}
