+++ docs/sortingalgorithms/README.md
@@ -1,0 +1,7 @@
+## Sorting Algorithms
+
+1.  [Selection Sort](SelectionSort)
+2.  [Merge Sort](MergeSort)
+3.  [Quick Sort](QuickSort)
+4.  [Move 0's to the End I](Move0sToTheEndI)
+5.  [Rainbow Sort](RainbowSort)
