+++ docs/sortingalgorithms/quick_sort1.png
