+++ docs/sortingalgorithms/quick_sort2.png
