+++ docs/sortingalgorithms/rainbowsort.png
