--- src/A/SortingAlgorithms/README.md
@@ -1,6 +1,0 @@
-## Search Algorithms
-1.  [Selection Sort](Easy/SelectionSort)
-2.  [Merge Sort](Medium/MergeSort)
-3.  [Quick Sort](Medium/QuickSort)
-4.  [Move 0's to the End I](Easy/Move0sToTheEndI)
-5.  [Rainbow Sort](Medium/RainbowSort)
