+++ tests/movezerostotheendi/Move0sToTheEndITest.java
@@ -1,0 +1,67 @@
+package tests.movezerostotheendi;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+
+class Move0sToTheEndITest {
+  static Move0sToTheEndI move0sToTheEndI;
+
+  @BeforeEach
+  void setUp() {
+    move0sToTheEndI = new Move0sToTheEndI();
+  }
+
+  @AfterEach
+  void tearDown() {}
+
+  @Test
+  void testNull() {
+    int[] result = move0sToTheEndI.moveZero(null);
+    assertArrayEquals(null, result);
+    int[] resultKeptOrder = move0sToTheEndI.moveZeroPreservePosition(null);
+    assertArrayEquals(null, resultKeptOrder);
+  }
+
+  @Test
+  void testEmpty() {
+    int[] result = move0sToTheEndI.moveZero(new int[] {});
+    assertArrayEquals(new int[] {}, result);
+    int[] resultKeptOrder = move0sToTheEndI.moveZeroPreservePosition(
+      new int[] {}
+    );
+    assertArrayEquals(new int[] {}, resultKeptOrder);
+  }
+
+  @Test
+  void testOne() {
+    int[] result = move0sToTheEndI.moveZero(new int[] { 1 });
+    assertArrayEquals(new int[] { 1 }, result);
+    int[] resultKeptOrder = move0sToTheEndI.moveZeroPreservePosition(
+      new int[] { 1 }
+    );
+    assertArrayEquals(new int[] { 1 }, resultKeptOrder);
+  }
+
+  @Test
+  void testTwo() {
+    int[] result = move0sToTheEndI.moveZero(new int[] { 1, 0, 3, 0, 1 });
+    assertArrayEquals(new int[] { 1, 1, 3, 0, 0 }, result);
+    int[] resultKeptOrder = move0sToTheEndI.moveZeroPreservePosition(
+      new int[] { 1, 0, 3, 0, 1 }
+    );
+    assertArrayEquals(new int[] { 1, 3, 1, 0, 0 }, resultKeptOrder);
+  }
+
+  @Test
+  void testThree() {
+    int[] result = move0sToTheEndI.moveZero(new int[] { 0, 0, 1, 1, 0, 1, 0 });
+    assertArrayEquals(new int[] { 1, 1, 1, 0, 0, 0, 0 }, result);
+    int[] resultKeptOrder = move0sToTheEndI.moveZeroPreservePosition(
+      new int[] { 0, 0, 1, 1, 0, 1, 0 }
+    );
+    assertArrayEquals(new int[] { 1, 1, 1, 0, 0, 0, 0 }, resultKeptOrder);
+  }
+}
